+++ nautilus_trader/data/engine.pyx
@@ -53,6 +53,9 @@ from nautilus_trader.core.data cimport Data
 from nautilus_trader.core.datetime cimport dt_to_unix_nanos
 from nautilus_trader.core.datetime cimport unix_nanos_to_dt
 from nautilus_trader.core.rust.common cimport ComponentState
+from nautilus_trader.core.rust.core cimport NANOSECONDS_IN_MILLISECOND
+from nautilus_trader.core.rust.core cimport NANOSECONDS_IN_SECOND
+from nautilus_trader.core.rust.core cimport millis_to_nanos
 from nautilus_trader.core.rust.model cimport PriceType
 from nautilus_trader.core.uuid cimport UUID4
 from nautilus_trader.data.aggregation cimport BarAggregator
@@ -799,18 +802,27 @@ cdef class DataEngine(Component):
             self._log.error("Cannot subscribe for synthetic instrument `OrderBook` data.")
             return
 
-        cdef int interval_ms = metadata["interval_ms"]
+        cdef:
+            uint64_t interval_ms = metadata["interval_ms"]
+            uint64_t interval_ns
+            uint64_t timestamp_ns
         key = (instrument_id, interval_ms)
         if key not in self._order_book_intervals:
             self._order_book_intervals[key] = []
-            now = self._clock.utc_now()
-            start_time = now - timedelta(milliseconds=int((now.second * 1000) % interval_ms), microseconds=now.microsecond)
-            timer_name = f"OrderBook_{instrument_id}_{interval_ms}"
-            self._clock.set_timer(
+
+            timer_name = f"OrderBook-{instrument_id}-{interval_ms}"
+            interval_ns = millis_to_nanos(interval_ms)
+            timestamp_ns = self._clock.timestamp_ns()
+            start_time_ns = timestamp_ns - (timestamp_ns % interval_ns)
+
+            if start_time_ns - NANOSECONDS_IN_MILLISECOND <= self._clock.timestamp_ns():
+                start_time_ns += NANOSECONDS_IN_SECOND  # Add one second
+
+            self._clock.set_timer_ns(
                 name=timer_name,
-                interval=timedelta(milliseconds=interval_ms),
-                start_time=start_time,
-                stop_time=None,
+                interval_ns=interval_ns,
+                start_time_ns=start_time_ns,
+                stop_time_ns=0,  # No stop
                 callback=self._snapshot_order_book,
             )
             self._log.debug(f"Set timer {timer_name}.")
@@ -1591,7 +1603,7 @@ cdef class DataEngine(Component):
         order_book.apply(data)
 
     cpdef void _snapshot_order_book(self, TimeEvent snap_event):
-        cdef tuple pieces = snap_event.name.partition('_')[2].partition('_')
+        cdef tuple pieces = snap_event.name.partition('-')[2].partition('-')
         cdef InstrumentId instrument_id = InstrumentId.from_str_c(pieces[0])
         cdef int interval_ms = int(pieces[2])